Some code rearranging and adding in fptp
