+++ election_dynamics/election_dynamics_multi_party_simple_voters.py
@@ -1,14 +1,8 @@
-from collections import Counter
-import copy
-import matplotlib.animation as animation
-import matplotlib.colors as mcolors
-import matplotlib.pyplot as plt
-import matplotlib.patches as patches
 import numpy as np
+from typing import Callable
 
 from election_dynamics.election_dynamics_multi_party import ElectionDynamicsMultiParty
 from policies.policy import Policy
-from utility_functions.evaluation_functions import ranked_choice_preference
 from voters.simple_voter import SimpleVoter
 
 
@@ -16,12 +10,13 @@ class ElectionDynamicsMultiPartySimpleVoters(ElectionDynamicsMultiParty):
     def __init__(
         self,
         voters: list[SimpleVoter],
+        evaluation_function: Callable,
         issue_1: str = "Issue 1",
         issue_2: str = "Issue 2",
     ):
         self.voters = voters
         self.voter_arr = np.array([voter.ideal_policy.values for voter in self.voters])
-        self.evaluation_function = ranked_choice_preference
+        self.evaluation_function = evaluation_function
         self.tiebreak_func = None
         self.issue_1 = issue_1  # Issue 1 name
         self.issue_2 = issue_2  # Issue 2 name
@@ -41,354 +36,3 @@ class ElectionDynamicsMultiPartySimpleVoters(ElectionDynamicsMultiParty):
         dists = np.linalg.norm(self.voter_arr[:, np.newaxis, :] - policies_arr[np.newaxis, :, :], axis=2)
         preferences = np.argsort(dists, axis=1)
         return preferences
-
-
-    def animate_election(
-        self, 
-        policies: list[Policy],
-        stop_at_majority: bool = True,
-        output_folder="output",
-        filename="rcv_election_animation",
-        plot_verbose=True,
-        fps=0.5,
-    ):
-        # setup
-        if len(policies) < 2:
-            print("Not enough policies to hold an election!")
-            return
-        if len(policies) > 10:
-            print("Not enough colors to adequately plot!")
-            return
-        
-        policies_arr = np.array([p.values for p in policies])
-        preferences = self.tabulate_votes(policies)
-        num_voters, num_candidates = preferences.shape
-        active = np.ones(num_candidates, dtype=bool)
-        done = {"stop": False}
-
-        original_first_choices = preferences[:, 0]
-        original_first_round_counts = np.bincount(original_first_choices, minlength=num_candidates)
-
-    
-        # color and name vis settings
-        mcolors_dict = copy.deepcopy(mcolors.TABLEAU_COLORS)
-
-        # ensuring blue and red are the first two colors used
-        del mcolors_dict['tab:blue']
-        del mcolors_dict['tab:red']
-        all_colors = ['blue', 'red'] + list(mcolors_dict.values())
-        policy_names = [policy.name for policy in policies]
-        policy_colors = all_colors[0:len(policies)]
-
-        # plot
-        # initialize the figure and axes
-        fig = plt.figure()
-
-        def make_frame(f_num):
-            # TODO: should this be subplots?
-            plt.clf()  # Clear the current axes/figure
-
-            mask = active[preferences]
-            top_choice_indices = mask.argmax(axis=1)
-            first_choices = preferences[np.arange(num_voters), top_choice_indices]
-
-            counts = np.bincount(first_choices, minlength=num_candidates)
-            total_active_votes = counts[active].sum()
-            ax = fig.add_axes([0.1, 0.3, 0.55, 0.55])  # Shrink plot inside the figure
-
-             # determining if the stopping condition has been met
-            for i in np.flatnonzero(active):
-                if (
-                    stop_at_majority and counts[i] > total_active_votes / 2
-                ) or (
-                    not stop_at_majority and len(active) == 2  and counts[i] > total_active_votes / 2
-                ):
-                    done["stop"] = True
-
-            # plotting all voters and policies
-            voters_by_vote = {}
-            for i in range(len(active)):
-                if active[i]:
-                    voters_by_vote[i] = []
-
-            for i, choice in enumerate(first_choices):
-                voters_by_vote[choice].append(self.voters[i])
-
-            for k in voters_by_vote.keys():
-                if active[k] and voters_by_vote[k]:
-                    arr = np.array(
-                        [voter.ideal_policy.values for voter in voters_by_vote[k]]
-                    )
-                    voters_plot = ax.scatter(
-                        arr[:, 0],
-                        arr[:, 1],
-                        c=policy_colors[k],
-                        marker="o",
-                    )
-                    curr_name = policy_names[k]
-                    voters_plot.set_label(f"{curr_name} Voters")
-
-                    policy_plot = ax.scatter(
-                        policies_arr[k,0],
-                        policies_arr[k,1],
-                        color=policy_colors[k],
-                        marker="o",
-                        edgecolors="black",
-                        s=200,
-                    )
-                    policy_plot.set_label(curr_name)
-
-            # title, labels, legend
-            desired_order = policy_names + [f"{policy_name} Voters" for policy_name in policy_names]
-            handles, labels = plt.gca().get_legend_handles_labels()
-            label_to_handle = dict(zip(labels, handles))
-            ordered_handles = [
-                label_to_handle[label]
-                for label in desired_order
-                if label in label_to_handle
-            ]
-            ordered_labels = [
-                label
-                for label in desired_order
-                if label in label_to_handle
-            ]
-            # reformatting ordered_labels to be shorter if too long
-            formatted_ordered_labels = []
-            for label in ordered_labels:
-                if len(label) > 18:
-                    formatted_ordered_labels.append(label[:18] + "...")
-                else:
-                    formatted_ordered_labels.append(label)
-
-            plt.legend(
-                loc="upper left",
-                bbox_to_anchor=(1.05, 1),
-                borderaxespad=0.0,
-                handles=ordered_handles,
-                labels=formatted_ordered_labels,
-            )
-
-            title = f"Ranked-choice Voting Round {f_num+1}\n"
-            too_many_policies = False
-            policy_added = False
-            for policy_name in policy_names:
-                if len(title) + len(str(policy_name)) > 80:
-                    too_many_policies = True
-                    break
-                else:
-                    title += f"{policy_name} vs. "
-                    policy_added = True
-            if too_many_policies and policy_added:
-                title += "..."
-            elif policy_added:
-                title = title[:-5]
-            plt.title(title)
-
-            plt.xlabel(f"Position on {self.issue_1}")
-            plt.ylabel(f"Position on {self.issue_2}")
-
-            if plot_verbose:
-                # sorting voters_by_vote by first round vote count
-                description_string = "In first-choice selections"
-                sorted_items = sorted(voters_by_vote.items(), key=lambda item: len(item[1]), reverse=True)
-                for i in range(len(sorted_items)):
-                    sorted_items[i] = list(sorted_items[i])
-                    sorted_items[i][1] = len(sorted_items[i][1])
-                top_policy_idx = sorted_items[0][0]
-                top_policy_name = policy_names[top_policy_idx]
-                top_policy_votes = sorted_items[0][1]
-                description_string = f"\n{top_policy_name} leads, as the first choice of {top_policy_votes} voters."
-
-                if len(sorted_items) > 1:
-                    second_policy_idx = sorted_items[1][0]
-                    second_policy_name = policy_names[second_policy_idx]
-                    second_policy_votes = sorted_items[1][1]
-                    if len(sorted_items) > 2:
-                        description_string += f"\n{second_policy_name} is in second place, as the first choice of {second_policy_votes} voters."
-                    else:
-                        # this is the last place candidate, phrase accordingly
-                        description_string += f"\n{second_policy_name} trails, as the first choice of {second_policy_votes} voters."
-                        description_string += f"\nSo, {top_policy_name} wins!"
-
-                if len(sorted_items) > 2:
-                    last_policy_idx = sorted_items[-1][0]
-                    last_policy_name = policy_names[last_policy_idx]
-                    last_policy_votes = sorted_items[-1][1]
-                    if len(sorted_items) > 3:
-                            description_string += "\n..."
-                    description_string += f"\n{last_policy_name} is in last place, as the first choice of {last_policy_votes} voters."
-                    if not done["stop"]:
-                        description_string += f"\nSo, {last_policy_name} is eliminated."
-                    else:
-                        # stopping condition has been met, declare winner
-                        description_string += f"\n{top_policy_name} has a majority of votes, so {top_policy_name} wins!"
-
-                fig.text(
-                    0.1,
-                    0.05,
-                    description_string,
-                    fontsize=8,
-                    color="black",
-                )
-            
-            # eliminating losers
-            min_votes = counts[active].min()
-            lowest = np.flatnonzero((counts == min_votes) & active)
-
-            if len(lowest) > 1:
-                orig_support = original_first_round_counts[lowest]
-                min_orig = orig_support.min()
-                lowest = lowest[orig_support == min_orig]
-
-            to_eliminate = lowest.min()
-            active[to_eliminate] = False
-
-        def init():
-            return None
-        
-        def frame_gen():
-            f_num = 0
-            while not done["stop"]:
-                if f_num >= len(policies) - 1:
-                    break
-                yield f_num
-                f_num += 1
-
-        ani = animation.FuncAnimation(
-            fig, make_frame, frames=frame_gen(), init_func=init, save_count=len(policies)+5  # leaving some extra buffer
-        )
-        # Save to mp4
-        ani.save(f"{output_folder}/{filename}.mp4", writer="ffmpeg", fps=fps)
-        print(f"Animation saved to {output_folder}/{filename}.png")
-        plt.close(fig)
-    
-
-    def create_election_sankey_diagram(
-        self, 
-        policies: list[Policy], 
-        stop_at_majority: bool = True,
-        output_folder="output",
-        filename=None
-    ):
-        """
-        Creates a round-by-round alluvial (Sankey-like) diagram using matplotlib, showing how votes are 
-        redistributed as candidates are eliminated. 
-        Each round is a vertical column (labeled at the top), each candidate is a horizontal row,
-        each node is labeled as policy_name and sized by vote count, and flows are drawn between nodes
-        in adjacent rounds.
-        Note: the sankey library is not used here, as its capabilities do not allow for the desired visualization.
-        """
-
-        if len(policies) < 2:
-            print("Not enough policies to hold an election!")
-            return
-        if len(policies) > 10:
-            print("Too many policies to plot properly!")
-            return
-        
-        preferences = self.tabulate_votes(policies)
-        num_voters, num_candidates = preferences.shape
-        policy_names = [p.name for p in policies]
-
-        # run RCV and collect round-by-round vote counts
-        vote_counts_by_round = self.evaluation_function(preferences, stop_at_majority=stop_at_majority, output_vote_counts=True)
-        num_rounds = vote_counts_by_round.shape[0]
-
-        # for each round, build a mapping from policy to y-position (for stacking)
-        y_positions = []
-        for r in range(num_rounds):
-            y = 0
-            pos = {}
-            for c in range(num_candidates):
-                if vote_counts_by_round[r, c] > 0:
-                    pos[c] = y
-                    y += vote_counts_by_round[r, c]
-            y_positions.append(pos)
-        total_votes = np.sum(vote_counts_by_round[0])
-
-        # prepare figure
-        fig, ax = plt.subplots(figsize=(9, 6))
-        ax.set_xlim(-0.5, num_rounds-0.5)
-        ax.set_ylim(0, total_votes)
-        ax.axis('off')
-
-        # draw round labels at the top
-        for r in range(num_rounds):
-            ax.text(r, total_votes + total_votes*0.04, f"Round {r+1}", ha='center', va='top', fontsize=11, fontweight='bold')
-
-        # draw candidate row labels on the left
-        for c in range(num_candidates):
-            # find the first round where candidate is present
-            for r in range(num_rounds):
-                if vote_counts_by_round[r, c] > 0:
-                    y = y_positions[r][c] + vote_counts_by_round[r, c]/2
-                    ax.text(-0.3, y, policy_names[c], ha='right', va='center', fontsize=11, fontweight='bold')
-                    break
-
-        # draw nodes (rectangles for each policy in each round)
-        node_width = 0.3
-        node_rects = {}
-        for r in range(num_rounds):
-            for c in range(num_candidates):
-                count = vote_counts_by_round[r, c]
-                if count > 0:
-                    y = y_positions[r][c]
-                    rect = patches.Rectangle((r-node_width/2, y), node_width, count, facecolor=f'C{c}', edgecolor='k', alpha=0.8)
-                    ax.add_patch(rect)
-                    node_rects[(c, r)] = (r-node_width/2, y, node_width, count)
-                    # Label
-                    ax.text(r, y+count/2, f"{count}", ha='center', va='center', fontsize=10, color='black')
-
-        # track which candidates are active each round
-        active = np.ones(num_candidates, dtype=bool)
-        current_preferences = preferences.copy()
-        for r in range(num_rounds-1):
-            counts = vote_counts_by_round[r]
-            next_counts = vote_counts_by_round[r+1]
-            # find eliminated candidate (active in this round, but not in next)
-            eliminated = np.where((counts > 0) & (next_counts == 0))[0]
-            if len(eliminated) == 0:
-                break  # No more eliminations
-            elim = eliminated[0]
-            # for each voter whose top active choice is elim, find their next active choice
-            mask = active[current_preferences]
-            top_choice_indices = mask.argmax(axis=1)
-            first_choices = current_preferences[np.arange(num_voters), top_choice_indices]
-            affected_voters = np.where(first_choices == elim)[0]
-            # for each affected voter, find their next preferred active candidate
-            transfer_targets = []
-            for v in affected_voters:
-                for pref in current_preferences[v]:
-                    if active[pref] and pref != elim:
-                        transfer_targets.append(pref)
-                        break
-            # count transfers
-            transfer_counter = Counter(transfer_targets)
-            # draw flows from elim in round r to each target in round r+1
-            y0 = y_positions[r][elim]
-            y1s = {c: y_positions[r+1][c] for c in transfer_counter}
-            offset0 = 0
-            offsets1 = {c: 0 for c in transfer_counter}
-            for c, n in transfer_counter.items():
-                # dource rectangle (elim, r): (r-node_width/2, y0+offset0, node_width, n)
-                # target rectangle (c, r+1): (r+1-node_width/2, y1s[c]+offsets1[c], node_width, n)
-                verts = [
-                    (r+node_width/2, y0+offset0),
-                    (r+1-node_width/2, y1s[c]+offsets1[c]),
-                    (r+1-node_width/2, y1s[c]+offsets1[c]+n),
-                    (r+node_width/2, y0+offset0+n)
-                ]
-                polygon = patches.Polygon(verts, closed=True, facecolor=f'C{elim}', edgecolor='none', alpha=0.4)
-                ax.add_patch(polygon)
-                offset0 += n
-                offsets1[c] += n
-            # eliminate the candidate
-            active[elim] = False
-        plt.title('RCV Vote Transfers (Sankey Diagram)', fontsize=16, pad=30, fontweight='bold')
-        plt.tight_layout()
-        if output_folder is not None and filename is not None:
-            plt.savefig(f"{output_folder}/{filename}.png", bbox_inches='tight')
-            print(f"Sankey diagram saved to {output_folder}/{filename}.png")
-        else:
-            plt.show()
-        plt.close(fig)
