+++ election_dynamics/electoral_systems.py
@@ -2,7 +2,7 @@ from election_dynamics.election_dynamics_multi_party_simple_voters import Electi
 from election_dynamics.election_dynamics_two_party import ElectionDynamicsTwoParty
 from election_dynamics.election_dynamics_two_party_simple_voters import ElectionDynamicsTwoPartySimpleVoters
 from election_dynamics.election_dynamics_two_party_taxicab_voters import ElectionDynamicsTwoPartyTaxicabVoters
-from utility_functions.evaluation_functions import status_quo_preference
+from utility_functions.evaluation_functions import first_past_the_post, ranked_choice_preference, status_quo_preference 
 from voters.simple_voter import SimpleVoter
 from voters.taxicab_voter import TaxicabVoter
 from voters.voter import Voter
@@ -53,11 +53,21 @@ def create_taxicab_electorate(
     )
 
 
+def create_fptp_multiparty_electorate(
+    voters: list[SimpleVoter],
+    issue_1: str = "Issue 1",
+    issue_2: str = "Issue 2",
+):
+    return ElectionDynamicsMultiPartySimpleVoters(
+        voters, evaluation_function=first_past_the_post, issue_1=issue_1, issue_2=issue_2,
+    )
+
+
 def create_rcv_electorate(
     voters: list[SimpleVoter],
     issue_1: str = "Issue 1",
     issue_2: str = "Issue 2",
 ):
     return ElectionDynamicsMultiPartySimpleVoters(
-        voters, issue_1=issue_1, issue_2=issue_2,
+        voters, evaluation_function=ranked_choice_preference, issue_1=issue_1, issue_2=issue_2,
     )
