+++ election_dynamics/electorates.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import numpy as np
 
-from election_dynamics.electoral_systems import create_rcv_electorate, create_simple_electorate
+from election_dynamics.electoral_systems import create_fptp_multiparty_electorate, create_rcv_electorate, create_simple_electorate
 from voters.simple_voter import SimpleVoter
 from voters.voter import Policy
 
@@ -32,9 +32,9 @@ def create_us_electorate_echelon_sample():
     return electorate
 
 
-def create_us_electorate_echelon_sample_rcv():
+def create_us_electorate_multiparty_rcv():
     """
-    Creates a RCV US electorate corresponding to the data from Echelon Insights, populated with Euclidean voters.
+    Creates a multiparty RCV US electorate corresponding to the data from Echelon Insights, populated with Euclidean voters.
     This electorate does not account for the voter weights present in Echelon's data, nor for other attributes in the
     data, outside of the Econ Scores and Social Scores of the voters.
 
@@ -55,6 +55,32 @@ def create_us_electorate_echelon_sample_rcv():
 
     # defining electorate
     electorate = create_rcv_electorate(voters, "Economic Conservatism", "Social Conservatism")
+    return electorate
+
+
+def create_us_electorate_echelon_multiparty_fptp():
+    """
+    Creates a multiparty FPTP US electorate corresponding to the data from Echelon Insights, populated with Euclidean voters.
+    This electorate does not account for the voter weights present in Echelon's data, nor for other attributes in the
+    data, outside of the Econ Scores and Social Scores of the voters.
+
+    Returns the created electorate.
+    """
+    # loading voter data
+    voter_data = pd.read_csv("data/us_electorate_echelon_sample_2025.csv")
+
+    # transforming and cleaning data
+    voter_data['Econ Score'] = pd.to_numeric(voter_data['Econ Score'], errors='coerce')
+    voter_data['Social Score'] = pd.to_numeric(voter_data['Social Score'], errors='coerce')
+    voter_data.dropna(subset=['Econ Score', 'Social Score'], inplace=True)
+
+    # defining voters
+    voters = [SimpleVoter(Policy(
+        np.array([row['Econ Score'], row['Social Score']])
+    )) for _, row in voter_data.iterrows()]
+
+    # defining electorate
+    electorate = create_fptp_multiparty_electorate(voters, "Economic Conservatism", "Social Conservatism")
     return electorate
